# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create an account with name, mobile number and password; signs in.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Sign in with mobile number and password.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "summary": "Exchange a refresh token for a new token pair (single use).",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sign out of this device.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sign out of every device.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "operationId": "AuthController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Devices currently signed in to this account.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions/{id}": {
      "delete": {
        "operationId": "AuthController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sign another device out.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "operationId": "DevicesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Phones and TVs signed in to this account.",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{id}": {
      "patch": {
        "operationId": "DevicesController_rename",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename a device.",
        "tags": [
          "devices"
        ]
      },
      "delete": {
        "operationId": "DevicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a device: it is signed out at once. Removing this device signs you out.",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/api/v1/pairing/sessions": {
      "post": {
        "operationId": "PairingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "TV: start pairing. Returns the code to show and a secret poll token.",
        "tags": [
          "pairing"
        ]
      }
    },
    "/api/v1/pairing/sessions/{id}/status": {
      "get": {
        "operationId": "PairingController_status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-pairing-poll-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "TV: poll. After approval the sign-in is returned exactly once.",
        "tags": [
          "pairing"
        ]
      }
    },
    "/api/v1/pairing/claim": {
      "post": {
        "operationId": "PairingController_claim",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Phone: enter or scan the code on the TV.",
        "tags": [
          "pairing"
        ]
      }
    },
    "/api/v1/pairing/{id}/approve": {
      "post": {
        "operationId": "PairingController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Phone: confirm \"Connect this TV?\".",
        "tags": [
          "pairing"
        ]
      }
    },
    "/api/v1/pairing/{id}/reject": {
      "post": {
        "operationId": "PairingController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Phone: \"Not my TV\".",
        "tags": [
          "pairing"
        ]
      }
    },
...
  },
  "info": {
    "title": "VideoBridge API",
    "description": "REST API for the VideoBridge phone and TV apps. Every non-2xx response uses the error envelope.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DeviceDto": {
        "type": "object",
        "properties": {
          "installId": {
            "type": "string",
            "description": "Random id created by the app on first launch."
          },
          "type": {
            "type": "string",
            "enum": [
              "PHONE",
              "ANDROID_TV"
            ]
          },
          "name": {
            "type": "string",
            "example": "OnePlus 13"
          },
          "manufacturer": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "osVersion": {
            "type": "string"
          },
          "appVersion": {
            "type": "string"
          }
        },
        "required": [
          "installId",
          "type",
          "name"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Mobile number; +91 is assumed without a country code."
          },
          "password": {
            "type": "string",
            "description": "Any password of 1 to 200 characters."
          },
          "device": {
            "$ref": "#/components/schemas/DeviceDto"
          },
          "name": {
            "type": "string",
            "example": "Janak"
          }
        },
        "required": [
          "phone",
          "password",
          "device",
          "name"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "phoneMasked": {
            "type": "string",
            "example": "+91******3210"
          },
          "displayName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "phoneMasked",
          "displayName",
          "createdAt"
        ]
      },
      "DeviceSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PHONE",
              "ANDROID_TV"
            ]
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "name"
        ]
      },
      "SessionRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceSummaryDto"
          },
          "session": {
            "$ref": "#/components/schemas/SessionRefDto"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "refreshToken",
          "refreshTokenExpiresAt",
          "user",
          "device",
          "session"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "9876543210",
            "description": "Mobile number; +91 is assumed without a country code."
          },
          "password": {
            "type": "string",
            "description": "Any password of 1 to 200 characters."
          },
          "device": {
            "$ref": "#/components/schemas/DeviceDto"
          }
        },
        "required": [
          "phone",
          "password",
          "device"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "refreshToken",
          "refreshTokenExpiresAt"
        ]
      },
      "SessionItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "device": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeviceSummaryDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "current": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "device",
          "createdAt",
          "lastUsedAt",
          "current"
        ]
      },
      "SessionListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "RenameDeviceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Living Room TV"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "TvDto": {
        "type": "object",
        "properties": {
          "installId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Living Room TV"
          },
          "type": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "osVersion": {
            "type": "string"
          },
          "appVersion": {
            "type": "string"
          }
        },
        "required": [
          "installId",
          "name"
        ]
      },
      "CreatePairingDto": {
        "type": "object",
        "properties": {
          "tv": {
            "$ref": "#/components/schemas/TvDto"
          }
        },
        "required": [
          "tv"
        ]
      },
      "ClaimDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "1234",
            "description": "The 4-digit code on the TV. Spaces and dashes are ignored."
          }
        },
        "required": [
          "code"
        ]
      },
      "ApproveDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Bedroom TV",
            "description": "Rename the TV while connecting it."
          }
        }
      },
...
    }
  }
}